Refactor inbound message source structure
