+++ package.json
@@ -32,7 +32,7 @@
     "cron": "^3.1.8",
     "ffc-messaging": "2.10.1",
     "hapi-pino": "12.1.0",
-    "joi": "17.13.3",
+    "joi": "^17.13.3",
     "log-timestamp": "0.3.0",
     "notifications-node-client": "8.2.1",
     "pg": "8.13.0",
