+++ test/unit/messages/inbound/comms-request/index.test.js
@@ -5,17 +5,17 @@ const mockReceiver = {
   abandonMessage: jest.fn()
 }
 
-jest.unstable_mockModule('../../../../app/messages/outbound/notification-status/index.js', () => ({
+jest.unstable_mockModule('../../../../../app/messages/outbound/notification-status/index.js', () => ({
   publishReceived: jest.fn()
 }))
 
-jest.unstable_mockModule('../../../../app/messages/inbound/send-notification.js', () => ({
+jest.unstable_mockModule('../../../../../app/messages/inbound/comms-request/send-notification.js', () => ({
   sendNotification: jest.fn()
 }))
 
-const { publishReceived } = await import('../../../../app/messages/outbound/notification-status/index.js')
-const { sendNotification } = await import('../../../../app/messages/inbound/send-notification.js')
-const { handleMessage } = await import('../../../../app/messages/inbound/handle-message.js')
+const { publishReceived } = await import('../../../../../app/messages/outbound/notification-status/index.js')
+const { sendNotification } = await import('../../../../../app/messages/inbound/comms-request/send-notification.js')
+const { handleCommsRequest } = await import('../../../../../app/messages/inbound/comms-request/index.js')
 
 describe('Handle Message', () => {
   beforeEach(() => {
@@ -25,7 +25,7 @@ describe('Handle Message', () => {
   test('should call publishReceived', async () => {
     const message = { body: 'mock-message' }
 
-    await handleMessage(message, mockReceiver)
+    await handleCommsRequest(message, mockReceiver)
 
     expect(publishReceived).toHaveBeenCalledWith(message.body)
   })
@@ -33,7 +33,7 @@ describe('Handle Message', () => {
   test('should call sendNotification', async () => {
     const message = { body: 'mock-message' }
 
-    await handleMessage(message, mockReceiver)
+    await handleCommsRequest(message, mockReceiver)
 
     expect(sendNotification).toHaveBeenCalledWith(message.body)
   })
@@ -41,7 +41,7 @@ describe('Handle Message', () => {
   test('should call completeMessage', async () => {
     const message = { body: 'mock-message' }
 
-    await handleMessage(message, mockReceiver)
+    await handleCommsRequest(message, mockReceiver)
 
     expect(mockReceiver.completeMessage).toHaveBeenCalledWith(message)
   })
@@ -51,7 +51,7 @@ describe('Handle Message', () => {
     const error = new Error('mock-error')
     publishReceived.mockRejectedValue(error)
 
-    await handleMessage(message, mockReceiver)
+    await handleCommsRequest(message, mockReceiver)
 
     expect(mockReceiver.completeMessage).not.toHaveBeenCalled()
     expect(mockReceiver.abandonMessage).toHaveBeenCalledWith(message)
@@ -62,7 +62,7 @@ describe('Handle Message', () => {
     const error = new Error('mock-error')
     sendNotification.mockRejectedValue(error)
 
-    await handleMessage(message, mockReceiver)
+    await handleCommsRequest(message, mockReceiver)
 
     expect(mockReceiver.completeMessage).not.toHaveBeenCalled()
     expect(mockReceiver.abandonMessage).toHaveBeenCalledWith(message)
@@ -73,7 +73,7 @@ describe('Handle Message', () => {
     const error = new Error('mock-error')
     mockReceiver.completeMessage.mockRejectedValue(error)
 
-    await handleMessage(message, mockReceiver)
+    await handleCommsRequest(message, mockReceiver)
 
     expect(mockReceiver.abandonMessage).toHaveBeenCalledWith(message)
   })
